Modify typeinfer tests
Description:
Currently the typeinfer test compares the gatetypes of gates with the expected types
by printing the bytecode name log and gatetype log. In order to pass the tests,
it relies on front end keeping the functions and bytecodes' order unchanged.
This is inappropriate and does not adapt to the increasing optimization of the front end.
Therefore, We need to take a more appropriate solution.
We plan to compare gatetypes of gates in Circuit IR and expected types at compile time.

Issue:
https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5HSKP

Change-Id: Id62a39ec0fc2327ede4e3ee1e770bc65bfc891d4
Signed-off-by: huoqingyi <huoqingyi@huawei.com>