Fix AOT do not compile methods from js files
A new bug was introduced from a recent commit. When the function
BytecodeInfoCollector::CollectInnerMethods is called, it is always
under the assumption that the abc file is obtained from one ts file.
This causes that if one abc file is obtained from one js file,
the inner-method information is not collected, resulting in
methods other than the main-method not being compiled by AOT.
The bug can be triggered when one abc file is obtained from one js file.

Issue:
https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5RJBG

Tests:
aot test, aot 262test

Change-Id: Ie4d97ac437975b5048e7162e74cea26b5aba16aa
Signed-off-by: huoqingyi <huoqingyi@huawei.com>