[Feature] Refactor the prototype object marking
Because the marking for prototype objects is not uniform, the IC is
impacted especailly for the IC about visiting the properties on
prototype chain. Therefore, we should unify the SetIsPrototype()
callsites to simplify the IC and improve the maintainability.

Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I90VHM

Testing: test262 and UT.

Signed-off-by: wuzhefeng <wuzhefeng1@huawei.com>
Change-Id: I8f37d77fc2663a5fa416bd6c5f4467c3acba09b1